add support for POST and PUT requests
The HttpRequest constructor takes two arguments:
- an ServerResponse instance
- a callback in which the request body is available
If is the property "data" of the instance NULL, there will be no calling of the callback. This is not a clean solution but we cannot call the callback right from the constructor because the instance would not be assigned to the variable in the original scope (app#_handleRequest).